refactor: Encapsulate number input sanitization
NumberTextFieldWrapper composable function encapsulates the logic for
sanitizing texfield's number input, allowing for reusing this logic in
different TextField UI components

+++ app/src/main/java/com/haidoan/android/stren/core/designsystem/component/Dialog.kt
@@ -157,7 +157,7 @@ fun AddMeasurementDialog(
                     modifier = Modifier.fillMaxWidth(),
                     number = inputValue,
                     onValueChange = {
-                        inputValue = it.toFloat()
+                        inputValue = it
                     },
                     label = "Value",
                     suffixText = selectedBiometrics.measurementUnit,
